consumer/requirements.txt
marshmallow==3.11.1
psycopg2-binary==2.8.6
black==20.8b1
green==3.2.5
kafka-python==2.0.2
pymitter==0.3.0
